+++ db/migrate/20220704184355_create_positions.rb
@@ -4,8 +4,8 @@ class CreatePositions < ActiveRecord::Migration[7.0]
       t.date :open_date
       t.string :symbol
       t.decimal :quantity
-      t.float :cost_per_share
-      t.float :commission_and_fee
+      t.decimal :cost_per_share
+      t.decimal :commission_and_fee
       t.references :portfolio, null: false, foreign_key: true
 
       t.timestamps
