+++ db/migrate/20220712161651_create_transactions.rb
@@ -5,9 +5,9 @@ class CreateTransactions < ActiveRecord::Migration[7.0]
       t.date :trade_date
       t.string :symbol
       t.integer :quantity
-      t.float :price
-      t.float :commission
-      t.float :fee
+      t.decimal :price
+      t.decimal :commission
+      t.decimal :fee
 
       t.timestamps
     end
